backend/track2_model/requirements.txt
# Core scientific libraries
numpy
scipy
pandas
matplotlib
tqdm

# Deep learning
torch>=2.6.0
torchvision
torchaudio
torchmetrics
# Transformers (CLIP, etc.)
transformers>=4.52.0
safetensors

# Image/Video processing
opencv-python
opencv-contrib-python
Pillow

# Dataset / ML helpers
scikit-learn
scikit-image

# MediaPipe (pose/hand extraction if you use it in preprocessing)
mediapipe

# Optional: GPU performance
accelerate

